added types and fixed
